Move higher-level API tests into a new test module
